app/app-requirements.txt
kafka-python
numpy
pandas
streamlit
plotly
